+++ sdk/textanalytics/azure-ai-textanalytics/src/test/java/com/azure/ai/textanalytics/TextAnalyticsClientTestBase.java
@@ -17,8 +17,8 @@ import com.azure.ai.textanalytics.models.ExtractKeyPhrasesOptions;
 import com.azure.ai.textanalytics.models.HealthcareEntity;
 import com.azure.ai.textanalytics.models.LinkedEntity;
 import com.azure.ai.textanalytics.models.LinkedEntityMatch;
-import com.azure.ai.textanalytics.models.SentenceOpinion;
 import com.azure.ai.textanalytics.models.PiiEntity;
+import com.azure.ai.textanalytics.models.PiiEntityCategory;
 import com.azure.ai.textanalytics.models.PiiEntityCollection;
 import com.azure.ai.textanalytics.models.PiiEntityDomainType;
 import com.azure.ai.textanalytics.models.RecognizeEntitiesActionResult;
@@ -26,6 +26,7 @@ import com.azure.ai.textanalytics.models.RecognizeEntitiesOptions;
 import com.azure.ai.textanalytics.models.RecognizeLinkedEntitiesOptions;
 import com.azure.ai.textanalytics.models.RecognizePiiEntitiesActionResult;
 import com.azure.ai.textanalytics.models.RecognizePiiEntitiesOptions;
+import com.azure.ai.textanalytics.models.SentenceOpinion;
 import com.azure.ai.textanalytics.models.SentenceSentiment;
 import com.azure.ai.textanalytics.models.TargetSentiment;
 import com.azure.ai.textanalytics.models.TextAnalyticsActions;
@@ -316,6 +317,10 @@ public abstract class TextAnalyticsClientTestBase extends TestBase {
 
     @Test
     abstract void recognizePiiEntitiesForBatchInputForDomainFilter(HttpClient httpClient,
+        TextAnalyticsServiceVersion serviceVersion);
+
+    @Test
+    abstract void recognizePiiEntitiesForBatchInputForCategoriesFilter(HttpClient httpClient,
         TextAnalyticsServiceVersion serviceVersion);
 
     // Linked Entities
@@ -622,6 +627,13 @@ public abstract class TextAnalyticsClientTestBase extends TestBase {
     @Test
     abstract void analyzeBatchActionsAllFailed(HttpClient httpClient, TextAnalyticsServiceVersion serviceVersion);
 
+    @Test
+    abstract void analyzePiiEntityRecognitionWithCategoriesFilters(HttpClient httpClient,
+        TextAnalyticsServiceVersion serviceVersion);
+
+    @Test
+    abstract void analyzeLinkedEntityTasks(HttpClient httpClient, TextAnalyticsServiceVersion serviceVersion);
+
     // Detect Language runner
     void detectLanguageShowStatisticsRunner(BiConsumer<List<DetectLanguageInput>,
         TextAnalyticsRequestOptions> testRunner) {
@@ -758,6 +770,13 @@ public abstract class TextAnalyticsClientTestBase extends TestBase {
     void recognizeStringBatchPiiEntitiesShowStatsRunner(
         BiConsumer<List<String>, RecognizePiiEntitiesOptions> testRunner) {
         testRunner.accept(PII_ENTITY_INPUTS, new RecognizePiiEntitiesOptions().setIncludeStatistics(true));
+    }
+
+    void recognizeStringBatchPiiEntitiesForCategoriesFilterRunner(
+        BiConsumer<List<String>, RecognizePiiEntitiesOptions> testRunner) {
+        testRunner.accept(PII_ENTITY_INPUTS,
+            new RecognizePiiEntitiesOptions().setCategoriesFilter(
+                PiiEntityCategory.USSOCIAL_SECURITY_NUMBER, PiiEntityCategory.ABAROUTING_NUMBER));
     }
 
     // Linked Entity runner
@@ -1034,6 +1053,28 @@ public abstract class TextAnalyticsClientTestBase extends TestBase {
                 .setRecognizePiiEntitiesOptions(
                     new RecognizePiiEntitiesOptions().setModelVersion("invalidaVersion"),
                     new RecognizePiiEntitiesOptions().setModelVersion("2929")));
+    }
+
+    void analyzePiiEntityRecognitionWithCategoriesFiltersRunner(
+        BiConsumer<List<TextDocumentInput>, TextAnalyticsActions> testRunner) {
+        testRunner.accept(
+            asList(
+                new TextDocumentInput("0", PII_ENTITY_INPUTS.get(0)),
+                new TextDocumentInput("1", PII_ENTITY_INPUTS.get(1))),
+            new TextAnalyticsActions()
+                .setDisplayName("Test1")
+                .setRecognizePiiEntitiesOptions(
+                    new RecognizePiiEntitiesOptions()
+                        .setCategoriesFilter(PiiEntityCategory.USSOCIAL_SECURITY_NUMBER)
+                ));
+    }
+
+    void analyzeLinkedEntityRecognitionRunner(BiConsumer<List<String>, TextAnalyticsActions> testRunner) {
+        testRunner.accept(
+            LINKED_ENTITY_INPUTS,
+            new TextAnalyticsActions()
+                .setDisplayName("Test1")
+                .setRecognizeLinkedEntitiesOptions(new RecognizeLinkedEntitiesOptions()));
     }
 
     String getEndpoint() {